FossilOrigin-Name: 0a4c8b093c649e7dcec21260e69737de9604e47b771fa6bdb8cd1f8bcb0de90f

+++ src/build.c
@@ -2085,8 +2085,8 @@ void sqlite3ChangeCookie(Parse *pParse, int iDb){
 ** The estimate is conservative.  It might be larger that what is
 ** really needed.
 */
-static int identLength(const char *z){
-  int n;
+static i64 identLength(const char *z){
+  i64 n;
   for(n=0; *z; n++, z++){
     if( *z=='"' ){ n++; }
   }
